ver 2.0 (final?) add stats level system and endless gamemode

+++ Project/Assets/Scripts/GeneralGame/Entities/StatsSystem/DamageArgs.cs
@@ -7,8 +7,8 @@ namespace Assets.Scripts.GeneralGame.Entities.StatsSystem
     public class DamageArgs
     {
         [SerializeField]
-        int damage = 0;
-        public int Damage { get => damage; set => damage = value; }
+        float damage = 0;
+        public float Damage { get => damage; set => damage = value; }
 
     }
 }
